added RP.X cut to both ana and simu
debugged behaviour of eff with RP.X

-> there is clearly a mismatch between ana and simu despite RP.X cut
since the eff peaks at large RP.X when our data starts decreasing there

->option: sample rp.x from analysis histograms, but there is a strong dependence
on the Ex... what is next?

+++ PostAnalysis/Pipelines/Pipe0_Beam.cxx
@@ -9,10 +9,6 @@
 
 #include <atomic>
 
-#include "../Gates.cxx"
-#include "../HistConfig.h"
-#include "../Utils.cxx"
-
 void Pipe0_Beam()
 {
     ROOT::EnableImplicitMT();
@@ -29,7 +25,7 @@ void Pipe0_Beam()
     auto hGATCONF {def.Histo1D("GATCONF")};
 
     // And cound CFA triggers
-    std::atomic<unsigned int> cfa {};
+    std::atomic<unsigned long int> cfa {};
     def.Foreach(
         [&](ActRoot::ModularData& mod)
         {
